+++ dynamic-neural-field-composer/include/user_interface/element_window.h
@@ -1,5 +1,9 @@
 #pragma once
 
+
+#include <unordered_map>
+#include <cmath>
+
 #include <imgui-platform-kit/user_interface_window.h>
 
 #include "simulation/simulation.h"
@@ -51,7 +55,8 @@ namespace dnf_composer::user_interface
 		static void setFocusedElement(const std::shared_ptr<element::Element>& element);
 		~ElementWindow() override = default;
 	private:
-		static void modifyElementNeuralField(const std::shared_ptr<element::Element>& element) ;
+		void renderDimensionControls(const std::shared_ptr<element::Element>& element) const;
+		static void modifyElementNeuralField(const std::shared_ptr<element::Element>& element);
 		static void modifyElementGaussStimulus(const std::shared_ptr<element::Element>& element);
 		static void modifyElementFieldCoupling(const std::shared_ptr<element::Element>& element);
 		static void modifyElementGaussKernel(const std::shared_ptr<element::Element>& element);